inline asm tests renaming
