+++ dartagnan/src/test/java/com/dat3m/dartagnan/inlineAsm/armv7/libvsync/AsmLibvsyncArmv7Test.java
@@ -1,4 +1,4 @@
-package com.dat3m.dartagnan.inlineAsm.armv8.ck;
+package com.dat3m.dartagnan.inlineAsm.armv7.libvsync;
 
 import java.io.File;
 import java.io.IOException;
@@ -26,21 +26,21 @@ import com.dat3m.dartagnan.program.Program;
 import static com.dat3m.dartagnan.utils.ResourceHelper.getRootPath;
 import static com.dat3m.dartagnan.utils.ResourceHelper.getTestResourcePath;
 import com.dat3m.dartagnan.utils.Result;
+import static com.dat3m.dartagnan.utils.Result.PASS;
 import com.dat3m.dartagnan.verification.VerificationTask;
-import com.dat3m.dartagnan.verification.solving.AssumeSolver;
 import com.dat3m.dartagnan.verification.solving.RefinementSolver;
 import com.dat3m.dartagnan.wmm.Wmm;
 
 @RunWith(Parameterized.class)
-public class InlineAsmTestArmv8Ck {
+public class AsmLibvsyncArmv7Test {
 
-    private final String modelPath = getRootPath("cat/aarch64.cat");
+    private final String modelPath = getRootPath("cat/arm.cat");
     private final String programPath;
     private final int bound;
     private final Result expected;
 
-    public InlineAsmTestArmv8Ck(String file, int bound, Result expected) {
-        this.programPath = getTestResourcePath("inlineasm/armv8/ck/" + file + ".ll");
+    public AsmLibvsyncArmv7Test(String file, int bound, Result expected) {
+        this.programPath = getTestResourcePath("inlineasm/armv7/libvsync/" + file + ".ll");
         this.bound = bound;
         this.expected = expected;
     }
@@ -48,32 +48,40 @@ public class InlineAsmTestArmv8Ck {
     @Parameterized.Parameters(name = "{index}: {0}, {1}, {2}")
     public static Iterable<Object[]> data() throws IOException {
         return Arrays.asList(new Object[][]{
-            {"anderson", 3, Result.PASS},
-            {"caslock", 3, Result.PASS},
-            {"clhlock", 1, Result.PASS},
-            {"declock", 3, Result.PASS},
-            {"faslock", 3, Result.PASS},
-            {"mcslock", 2, Result.PASS},
-            {"ticketlock", 1, Result.PASS},
-            {"spsc_queue", 1, Result.PASS},
-            {"stack_empty", 1, Result.PASS},
+            //bounded_queue
+            {"bounded_spsc", 4, PASS},
+            {"bounded_mpmc_check_full", 5, PASS},
+            {"bounded_mpmc_check_empty", 4, PASS},
+
+            //spinlocks
+            // {"caslock", 4, PASS}, // passes Refinement but out of memory on Assume 
+            {"clhlock", 3, PASS},
+            // {"cnalock", 5, PASS}, // takes 35 minutes
+            {"hemlock", 3, PASS},
+            {"mcslock", 3, PASS},
+            {"rec_mcslock", 3, PASS},
+            // {"rec_seqlock", 3, PASS}, // 25 min to pass
+            {"rec_spinlock", 3, PASS},
+            {"rwlock", 3, PASS},
+            {"semaphore", 3, PASS},
+            {"seqcount", 1, PASS},
+            {"seqlock", 3, PASS},
+            {"ttaslock", 3, PASS},
+            {"twalock", 2, PASS},
+
+            //threads 
+            {"mutex_musl", 3, PASS},
+            {"mutex_slim", 3, PASS},
+            {"mutex_waiters", 3, PASS},
+            {"once", 5, PASS}
         });
     }
 
     @Test
     public void testAllSolvers() throws Exception {
-        long start = System.currentTimeMillis();
-        System.out.println("\n " + this.programPath);
         try (SolverContext ctx = mkCtx(); ProverWithTracker prover = mkProver(ctx)) {
             assertEquals(expected, RefinementSolver.run(ctx, prover, mkTask()).getResult());
         }
-        
-        System.out.println("\n" + (System.currentTimeMillis() - start) + " time elapsed Refinment for " + this.programPath);
-        start = System.currentTimeMillis();
-        try (SolverContext ctx = mkCtx(); ProverWithTracker prover = mkProver(ctx)) {
-            assertEquals(expected, AssumeSolver.run(ctx, prover, mkTask()).getResult());
-        }
-        System.out.println("\n" + (System.currentTimeMillis() - start) + " time elapsed Assume for " + this.programPath);
     }
 
     private SolverContext mkCtx() throws InvalidConfigurationException {
@@ -93,7 +101,7 @@ public class InlineAsmTestArmv8Ck {
         VerificationTask.VerificationTaskBuilder builder = VerificationTask.builder()
                 .withConfig(Configuration.builder().build())
                 .withBound(bound)
-                .withTarget(Arch.ARM8);
+                .withTarget(Arch.ARM7);
         Program program = new ProgramParser().parse(new File(programPath));
         Wmm mcm = new ParserCat().parse(new File(modelPath));
         return builder.build(program, mcm, EnumSet.of(LIVENESS, PROGRAM_SPEC));
